fix: escape HTML sinks and harden pre-release helpers
Escape AccessibleCard, Button attrs, GetTerms, and Breadcrumb link
text; tighten PresetColor hex validation; fail closed on meta-gated
blocks without a post ID; fix BusinessHours cross-year/closed schema
and i18n domains. Expand Breadcrumbs and related unit coverage.

+++ inc/Utilities/ReadingTime.php
@@ -28,8 +28,8 @@ class ReadingTime
 		 * @param int $words_per_minute Words per minute. Default 200.
 		 */
 		$words_per_minute = apply_filters('wp_utility_reading_time_wpm', 200);
-		
-		$readingtime = ceil($word_count / $words_per_minute);
+
+		$readingtime = (int) ceil($word_count / $words_per_minute);
 		return $readingtime;
 	}
 }
